fork: stop suppressing warnings that were hiding real diagnostics
The build ran -Wall -Wextra with five blanket -Wno- flags. Measured one at a
time, four were hiding nothing of ours: -Wunused-value and -Wpointer-arith fired
zero times, -Wnarrowing three, and -Wunused-parameter 83 - of which 75 came from
hyprland's own headers.

Header noise is what SYSTEM on the hyprland include directories is for. With
that, the same diagnostics apply to our code and stay quiet about theirs, and
four of the five flags are simply gone.

Fixed on the way through:

  - three narrowing conversions, two int->double CBox fields and a long->int
    rounding radius. Benign, but -Wnarrowing is worth having in code that does
    this much geometry arithmetic.
  - eight unused parameters. Six are signatures the lua and dispatcher APIs
    dictate and are now unnamed; `silent` in moveTargetInDirection is likewise
    an interface parameter with nothing to do.
  - `from` in findTiledWindowCandidate, which is the one worth noting: it went
    dead when the workspace parameters were dropped a few commits ago, and
    nothing said so until this warning came back on. It is gone, and the
    asymmetry with findFloatingWindowCandidate - which really does search from a
    given window's workspace - is now the honest shape.

-Wmissing-field-initializers stays, and only it. 203 of its 208 were hyprland's
ColorManagement.hpp and went away with SYSTEM; the five left are one deliberate
C++20 designated initializer, which is the construct the warning cannot
distinguish from a mistake.

Clean build is warning-free. Suite 106/0.

+++ src/TabGroup.cpp
@@ -364,8 +364,8 @@ void Hy3TabBarEntry::renderText(float scale, CBox& box, float opacity) {
 	auto texture_box = CBox {
 	    box.x + x_offset + this->last_render.texture_x_offset,
 	    box.y + y_offset + this->last_render.texture_y_offset,
-	    this->last_render.texture_width,
-	    this->last_render.texture_height,
+	    static_cast<double>(this->last_render.texture_width),
+	    static_cast<double>(this->last_render.texture_height),
 	};
 
 	texture_box.round();
@@ -848,7 +848,7 @@ void Hy3TabGroup::renderTabBar() {
 			window_box.scale(scale);
 
 			if (window_box.width > 0 && window_box.height > 0)
-				Render::GL::g_pHyprOpenGL->renderRect(window_box, CHyprColor(0, 0, 0, 0), { .round = radius });
+				Render::GL::g_pHyprOpenGL->renderRect(window_box, CHyprColor(0, 0, 0, 0), { .round = static_cast<int>(radius) });
 		}
 
 		glColorMask(GL_TRUE, GL_TRUE, GL_TRUE, GL_TRUE);
